refactor(filemanager): 拆出类型契约与 ChmodDialog/RenameInput/ContextMenu
- 新增 components/filemanager/：fileManagerTypes（共享类型与 props 契约）、
  ChmodDialog（修改权限对话框）、RenameInput（行内重命名输入框）、
  ContextMenu（右键菜单，映射为 ui/ContextMenu items）
- FileManager.tsx 7120 -> 6547 行
