+++ frontend/src/components/FileManager.tsx
@@ -43,603 +43,28 @@ import {
   MonitorSmartphone, PencilLine, FolderSymlink, FileSymlink,
   type LucideIcon,
 } from 'lucide-react';
-
-// ============================================================
-// FileManager 类型契约（props 见 FileManagerProps；内部数据模型见下）
-// ============================================================
-
-// 来自 Go 桥/事件/localStorage 持久化数据的外部形状：字段以守卫读取，
-// 均带索引签名（运行时宽松数据，新字段无需改接口）。
-
-/** 工作区持久化文件标签（extractManualPinnedTabsFromWorkspace 输入） */
-interface FileManagerWorkspaceTab {
-  pinned?: unknown
-  systemPinned?: unknown
-  id?: unknown
-  path?: unknown
-  customTitle?: unknown
-}
-
-/** 工作区对象（含 tabs 列表） */
-interface FileManagerWorkspace {
-  tabs?: FileManagerWorkspaceTab[]
-}
-
-/** 虚拟滚动行条目（paneState.rows 元素） */
-interface FileManagerVirtualRowEntry {
-  logicalPath?: unknown
-  rowType?: unknown
-  item?: unknown
-  [key: string]: unknown
-}
-
-/** 文件管理面板状态（paneState：虚拟滚动面板形状，见 renderFileManagerVirtualViewport 字面量） */
-interface FileManagerPaneStateLike {
-  rows: FileManagerVirtualRow[]
-  [key: string]: unknown
-}
-
-/** 虚拟滚动可见区间 */
-
-/** 下载冲突（PreviewDownloadConflicts 返回项 / buildDownloadConflictMessage 输入） */
-interface FileManagerDownloadConflict {
-  relativePath?: unknown
-  localSize?: unknown
-  remoteSize?: unknown
-  localModifyTime?: unknown
-  remoteModifyTime?: unknown
-  localPath?: unknown
-  remotePath?: unknown
-  [key: string]: unknown
-}
-
-/** 下载冲突解决设置（buildDownloadConflictOptionsPayload 输入） */
-
-/** 文件项宽松形状（handleOpenSystemEditor/WithEditor 输入） */
-interface FileManagerFileLike {
-  name?: unknown
-  path?: unknown
-  content?: unknown
-  size?: unknown
-  permission?: unknown
-  mode?: unknown
-  uid?: unknown
-  gid?: unknown
-}
-
-/** 双面板拖拽 payload 项 */
-interface FileManagerDualPaneDragItem {
-  path?: unknown
-  isDirectory?: unknown
-  [key: string]: unknown
-}
-
-/** 双面板拖拽 payload（confirmDualPaneDirectoryDrag 输入） */
-interface FileManagerDualPaneDragPayload {
-  items?: FileManagerDualPaneDragItem[]
-  paths?: unknown
-  [key: string]: unknown
-}
-
-/** 标签拖放指示器（setFileManagerTabDropIndicator 状态） */
-interface FileManagerTabDropIndicator {
-  tabId?: unknown
-  [key: string]: unknown
-}
-
-/** chmod 目标（setChmodTarget 状态，{ item, path, mode, includeSubdirectories, showIncludeSubdirectories }） */
-interface FileManagerChmodTarget {
-  item: FileManagerFileItem | null
-  path: string
-  mode: string
-  includeSubdirectories?: boolean
-  showIncludeSubdirectories?: boolean
-  rememberedMode?: string
-  autoApplyLastSettings?: boolean
-  ownerCandidates?: IdentityPresetOption[]
-  groupCandidates?: IdentityPresetOption[]
-  [key: string]: unknown
-}
-
-interface FileManagerProps {
-  sessionId: string
-  sessionGroupId: string
-  addToast?: (message: string | Error, type?: string, duration?: number, actions?: unknown[]) => number
-  isActive?: boolean
-  initialPath?: string
-}
-
-// 远端/本地文件条目（ListDir 返回项 + 本地传输占位项的统一形状）
-interface FileManagerFileItem {
-  name: string
-  isDirectory: boolean
-  size?: number
-  permission?: string
-  mode?: string
-  modifyTime?: number
-  uid?: string
-  gid?: string
-  isSymlink?: boolean
-  __rowKey?: string
-  __luminDeletedPlaceholder?: boolean
-  [key: string]: unknown
-}
-
-
-
-
-
-
-type LooseT = (key: I18nKey, vars?: Record<string, unknown>) => string
-
-// 面板选中态恢复（FileManagerPaneState + 最后点击路径）
-interface PaneSelectionRestore extends FileManagerPaneState {
-  lastClickedPath?: string | null
-}
-
-// syncCurrentTabToWorkspace 的覆盖参数
-interface SyncTabOverrides {
-  path?: string
-  sortField?: string
-  sortDir?: string
-  selectedPaths?: string[]
-  scrollTop?: number
-  reason?: string
-}
-
-// 右键菜单状态（contextMenu state）
-interface ContextMenuState {
-  pos: { x: number; y: number }
-  item: FileManagerFileItem | null
-  mode?: string
-  itemBasePath?: string
-  createBasePath?: string
-  showCreateActions?: boolean
-  deleteItemCount?: number
-  clipboardItemCount?: number
-  deleteUsesSelectedPaths?: boolean
-  clipboardUsesSelectedPaths?: boolean
-  tabId?: string
-  tabPath?: string
-  tabPinned?: boolean
-  tabSystemPinned?: boolean
-}
-
-// loadDir 的选项
-interface LoadDirOptions {
-  silent?: boolean
-  tabId?: string
-  staleWhileRevalidate?: boolean
-  staleItems?: FileManagerFileItem[]
-  preferPathCache?: boolean
-  preserveWorkspacePathOnSuccess?: boolean
-  preserveView?: boolean
-  trackDiff?: boolean
-  showLoading?: boolean
-}
-
-// 跨实例共享的剪贴板/编辑器状态（挂在 window 上，形状以运行时为准）
-declare global {
-  interface Window {
-    __luminClipboards?: Record<string, unknown>
-    __luminEditorStates?: Record<string, unknown>
-  }
-}
-
-interface ChmodDialogProps {
-  path: string
-  permission: string
-  mode: string
-  rememberedMode?: string
-  autoApplyLastSettings?: boolean
-  uid: string
-  gid: string
-  ownerCandidates?: IdentityPresetOption[]
-  groupCandidates?: IdentityPresetOption[]
-  includeSubdirectories?: boolean
-  showIncludeSubdirectories?: boolean
-  onSave: (mode: string, includeChildren: boolean, ownerInput: string, groupInput: string) => void
-  onClose: () => void
-  t: LooseT
-}
-
-interface RenameInputProps {
-  initialValue: string
-  isDirectory: boolean
-  onConfirm: (value: string, refocus: boolean) => void
-  onCancel: () => void
-  mountedRef?: React.MutableRefObject<HTMLInputElement | null>
-}
-
-interface ContextMenuProps {
-  pos: { x: number; y: number }
-  item: FileManagerFileItem | null
-  mode?: string
-  isPinned?: boolean
-  isSystemPinned?: boolean
-  canTogglePinned?: boolean
-  canCloseTab?: boolean
-  showCreateActions?: boolean
-  deleteItemCount?: number
-  clipboardItemCount?: number
-  canPaste?: boolean
-  clipboardActionArrow?: string
-  onClose: () => void
-  onDownload: () => void
-  onEdit: () => void
-  onOpenSystemEditor: () => void
-  onOpenWithEditor: () => void
-  onRename: () => void
-  onDelete: () => void
-  onDeleteShell: () => void
-  onMkdir: () => void
-  onNewFile: () => void
-  onCompress: () => void
-  onUncompress: () => void
-  onChmod: () => void
-  onCopyPath: () => void
-  onCopyItem: () => void
-  onCutItem: () => void
-  onPaste: () => void
-  onOpenInNewTab: () => void
-  onTogglePinned: () => void
-  onCloseTab: () => void
-  t: LooseT
-}
-
-// 格式化文件大小
-// ── Chmod Dialog ──────────────────────────────────────────────
-function ChmodDialog({ path, permission, mode, rememberedMode = '', autoApplyLastSettings = false, uid, gid, ownerCandidates = [], groupCandidates = [], includeSubdirectories = false, showIncludeSubdirectories = false, onSave, onClose, t }: ChmodDialogProps) {
-  const parsePerms = (permStr: unknown) => {
-    const normalizedPermStr = String(permStr || '');
-    const p = normalizedPermStr.length >= 10 ? normalizedPermStr.slice(1) : '---------';
-    return {
-      user: { r: p[0] === 'r', w: p[1] === 'w', x: p[2] === 'x' },
-      group: { r: p[3] === 'r', w: p[4] === 'w', x: p[5] === 'x' },
-      other: { r: p[6] === 'r', w: p[7] === 'w', x: p[8] === 'x' },
-    };
-  };
-
-  const currentMode = normalizeChmodMode(mode);
-  const lastMode = normalizeChmodMode(rememberedMode);
-  const initialMode = autoApplyLastSettings && lastMode ? lastMode : currentMode;
-  const fallbackPerms = parsePerms(permission || '');
-  const [perms, setPerms] = useState(initialMode ? permsFromChmodMode(initialMode) : fallbackPerms);
-  const [octal, setOctal] = useState(initialMode || calcChmodOctal(fallbackPerms));
-  const [includeChildren, setIncludeChildren] = useState(autoApplyLastSettings ? Boolean(includeSubdirectories) : false);
-  const ownerOptions = useMemo(() => buildIdentityOptionList(uid, ownerCandidates), [uid, ownerCandidates]);
-  const groupOptions = useMemo(() => buildIdentityOptionList(gid, groupCandidates), [gid, groupCandidates]);
-  const ownerDefaultValue = useMemo(() => resolveIdentityInputValue(uid, ownerCandidates), [uid, ownerCandidates]);
-  const groupDefaultValue = useMemo(() => resolveIdentityInputValue(gid, groupCandidates), [gid, groupCandidates]);
-  const [ownerInput, setOwnerInput] = useState(ownerDefaultValue);
-  const [groupInput, setGroupInput] = useState(groupDefaultValue);
-  const [ownerTouched, setOwnerTouched] = useState(false);
-  const [groupTouched, setGroupTouched] = useState(false);
-
-  useEffect(() => {
-    setOwnerTouched(false);
-  }, [path, uid]);
-
-  useEffect(() => {
-    setGroupTouched(false);
-  }, [path, gid]);
-
-  useEffect(() => {
-    if (!ownerTouched) {
-      setOwnerInput(ownerDefaultValue);
-    }
-  }, [ownerDefaultValue, ownerTouched]);
-
-  useEffect(() => {
-    if (!groupTouched) {
-      setGroupInput(groupDefaultValue);
-    }
-  }, [groupDefaultValue, groupTouched]);
-
-  const filteredOwnerOptions = useMemo(() => {
-    const query = String(ownerInput || '').trim().toLowerCase();
-    const candidates = query
-      ? ownerOptions.filter((option) => option.searchText.includes(query))
-      : ownerOptions;
-    return candidates.slice(0, 80);
-  }, [ownerInput, ownerOptions]);
-
-  const filteredGroupOptions = useMemo(() => {
-    const query = String(groupInput || '').trim().toLowerCase();
-    const candidates = query
-      ? groupOptions.filter((option) => option.searchText.includes(query))
-      : groupOptions;
-    return candidates.slice(0, 80);
-  }, [groupInput, groupOptions]);
-
-  const togglePerm = (cat: 'user' | 'group' | 'other', key: 'r' | 'w' | 'x') => {
-    setPerms(prev => {
-      const next = { ...prev, [cat]: { ...prev[cat], [key]: !prev[cat][key] } };
-      setOctal(calcChmodOctal(next));
-      return next;
-    });
-  };
-
-  const handleOctalChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const val = normalizeChmodMode(e.target.value);
-    setOctal(val);
-    if (val.length === 3) {
-      setPerms(permsFromChmodMode(val));
-    }
-  };
-
-  const canApplyLastSettings = Boolean(lastMode);
-  const handleApplyLastSettings = () => {
-    if (!lastMode) {
-      return;
-    }
-    setOctal(lastMode);
-    setPerms(permsFromChmodMode(lastMode));
-    setIncludeChildren(Boolean(includeSubdirectories));
-  };
-
-  if (typeof document === 'undefined') {
-    return null;
-  }
-  return createPortal(
-    <div className="modal-overlay" style={{ zIndex: Z.MODAL }}>
-      <div className="modal modal-sm">
-        <div className="modal-header">
-          <div className="modal-title"><Lock size={14} /> {t('修改权限')}</div>
-        </div>
-        <div className="modal-body">
-          <div className="chmod-dialog-body">
-            <div className="chmod-dialog-path">{path}</div>
-            <div className="grid gap-[10px] my-3">
-              <div className="grid gap-1.5">
-                <label htmlFor="chmod-owner-input" className="text-sm text-tertiary">{t('属主')}</label>
-                <input
-                  id="chmod-owner-input"
-                  name="chmod-owner"
-                  className="input"
-                  list="chmod-owner-options"
-                  autoComplete="off"
-                  value={ownerInput}
-                  onChange={(e) => {
-                    setOwnerTouched(true);
-                    setOwnerInput(e.target.value);
-                  }}
-                  placeholder={t('搜索或输入属主...')}
-                />
-              </div>
-              <datalist id="chmod-owner-options">
-                {filteredOwnerOptions.map((option) => (
-                  <option key={option.label} value={option.label} />
-                ))}
-              </datalist>
-              <div className="grid gap-1.5">
-                <label htmlFor="chmod-group-input" className="text-sm text-tertiary">{t('属组')}</label>
-                <input
-                  id="chmod-group-input"
-                  name="chmod-group"
-                  className="input"
-                  list="chmod-group-options"
-                  autoComplete="off"
-                  value={groupInput}
-                  onChange={(e) => {
-                    setGroupTouched(true);
-                    setGroupInput(e.target.value);
-                  }}
-                  placeholder={t('搜索或输入属组...')}
-                />
-              </div>
-              <datalist id="chmod-group-options">
-                {filteredGroupOptions.map((option) => (
-                  <option key={option.label} value={option.label} />
-                ))}
-              </datalist>
-            </div>
-            <div className="chmod-grid">
-              <div className="chmod-row">
-                <span></span>
-                <span className="text-center text-sm text-tertiary">{t('读取')}</span>
-                <span className="text-center text-sm text-tertiary">{t('写入')}</span>
-                <span className="text-center text-sm text-tertiary">{t('执行')}</span>
-              </div>
-              <div className="chmod-row">
-                <span className="chmod-row-label">{t('用户')}</span>
-                {(['r','w','x'] as const).map(k => (
-                  <label key={k} htmlFor={`fm-chmod-user-${k}`} className="chmod-checkbox" style={{ justifyContent: 'center' }}>
-                    <input type="checkbox" id={`fm-chmod-user-${k}`} name={`fm-chmod-user-${k}`} autoComplete="off" checked={perms.user[k]} onChange={() => togglePerm('user', k)} />
-                  </label>
-                ))}
-              </div>
-              <div className="chmod-row">
-                <span className="chmod-row-label">{t('组')}</span>
-                {(['r','w','x'] as const).map(k => (
-                  <label key={k} htmlFor={`fm-chmod-group-${k}`} className="chmod-checkbox" style={{ justifyContent: 'center' }}>
-                    <input type="checkbox" id={`fm-chmod-group-${k}`} name={`fm-chmod-group-${k}`} autoComplete="off" checked={perms.group[k]} onChange={() => togglePerm('group', k)} />
-                  </label>
-                ))}
-              </div>
-              <div className="chmod-row">
-                <span className="chmod-row-label">{t('其他')}</span>
-                {(['r','w','x'] as const).map(k => (
-                  <label key={k} htmlFor={`fm-chmod-other-${k}`} className="chmod-checkbox" style={{ justifyContent: 'center' }}>
-                    <input type="checkbox" id={`fm-chmod-other-${k}`} name={`fm-chmod-other-${k}`} autoComplete="off" checked={perms.other[k]} onChange={() => togglePerm('other', k)} />
-                  </label>
-                ))}
-              </div>
-            </div>
-            <div className="flex items-center gap-3">
-              <label htmlFor="chmod-octal-input" className="text-sm text-tertiary">{t('八进制:')}</label>
-              <input id="chmod-octal-input" name="chmod-octal" className="chmod-octal-input" autoComplete="off" value={octal} onChange={handleOctalChange} />
-              <Button variant="ghost" size="sm" onClick={handleApplyLastSettings} disabled={!canApplyLastSettings}>
-                {t('应用上次')}
-              </Button>
-            </div>
-            {showIncludeSubdirectories && (
-              <label htmlFor="fm-chmod-include-children" className="chmod-checkbox" style={{ display: 'flex', alignItems: 'center', gap: 8, marginTop: 12 }}>
-                <input type="checkbox" id="fm-chmod-include-children" name="fm-chmod-include-children" autoComplete="off" checked={includeChildren} onChange={(e) => setIncludeChildren(e.target.checked)} />
-                <span>{t('包含子目录')}</span>
-              </label>
-            )}
-          </div>
-        </div>
-        <div className="modal-footer">
-          <Button variant="ghost" onClick={onClose}>{t('取消')}</Button>
-          <Button variant="primary" onClick={() => onSave(octal.length === 3 ? octal : calcChmodOctal(perms), includeChildren, ownerInput, groupInput)}>{t('确定')}</Button>
-        </div>
-      </div>
-    </div>,
-    document.body
-  );
-}
-
-// 行内重命名输入框。
-//
-// 设计要点：
-//   - 非受控（defaultValue + ref 读值），避免受控渲染竞态丢字符。
-//   - suppressDragOutClick 抑制"框内 mousedown 拖出 mouseup"派生的 click，
-//     防止冒泡到行级 onClick 抢焦点、触发 onBlur 误提交。
-//   - committedRef 保证 onBlur / Enter / 外部取消 三条提交路径只生效一次。
-//   - 虚拟化卸载（Virtuoso 把该行滚出视口）时 React 不触发 onBlur，renamingItem
-//     会残留——卸载后 cleanup 故意保留已脱离 DOM 的元素引用，
-//     由 F2 入口（handleFileListKeyDown）与行级点击路径检测“input 已脱离 DOM”
-//     并读取 .value 兜底提交。之所以不用 useEffect cleanup 提交：
-//     StrictMode 下会 mount→unmount→remount，cleanup 会在用户什么都没做时就误触发提交。
-function RenameInput({ initialValue, isDirectory, onConfirm, onCancel, mountedRef }: RenameInputProps) {
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const committedRef = useRef(false);
-
-  const commit = useCallback((refocus: boolean) => {
-    if (committedRef.current) return;
-    committedRef.current = true;
-    const el = inputRef.current;
-    const value = el ? el.value : '';
-    onConfirm(value, refocus);
-  }, [onConfirm]);
-
-  useEffect(() => {
-    const el = inputRef.current;
-    if (!el) return;
-    if (mountedRef) mountedRef.current = el; // 登记：供残留检测路径判断 input 是否还在 DOM
-    const name = el.value;
-    const extensionIndex = isDirectory ? -1 : name.lastIndexOf('.');
-    const selectionEnd = extensionIndex > 0 && extensionIndex < name.length - 1 ? extensionIndex : name.length;
-    el.setSelectionRange(0, selectionEnd);
-    // 卸载时不清空 mountedRef：保留已脱离 DOM 的元素引用，
-    // 让残留清理路径（F2 / 行级点击）能读到 .value 提交用户最后输入；
-    // 读取方都必须先做 document.body.contains 判断。绝不在 cleanup 做业务提交。
-  }, [isDirectory, mountedRef]);
-
-  return (
-    <input
-      ref={inputRef}
-      id="fm-rename-input"
-      name="fm-rename-input"
-      className="rename-input"
-      autoComplete="off"
-      defaultValue={initialValue}
-      onMouseDown={(event) => suppressDragOutClick(event.nativeEvent)}
-      onBlur={() => commit(false)}
-      onKeyDown={(event) => {
-        event.stopPropagation();
-        if (event.key === 'Enter') commit(true);
-        if (event.key === 'Escape') {
-          if (committedRef.current) return;
-          committedRef.current = true;
-          onCancel();
-        }
-      }}
-      onClick={(event) => event.stopPropagation()}
-      autoFocus
-    />
-  );
-}
-
-// Context menu component（视觉迁移：手写 DOM 换 ui/ContextMenu，业务条件原样映射为 items）
-function ContextMenu({ pos, item, mode = 'item', isPinned = false, isSystemPinned = false, canTogglePinned = false, canCloseTab = false, showCreateActions = false, deleteItemCount = 1, clipboardItemCount = 1, canPaste = false, clipboardActionArrow = '', onClose, onDownload, onEdit, onOpenSystemEditor, onOpenWithEditor, onRename, onDelete, onMkdir, onNewFile, onCompress, onUncompress, onChmod, onCopyPath, onCopyItem, onCutItem, onPaste, onOpenInNewTab, onTogglePinned, onCloseTab, t }: ContextMenuProps) {
-  const isTabMenu = mode === 'tab';
-  const shouldShowCreateActions = showCreateActions || !item;
-  const shouldShowDividerBeforeCreate = Boolean(item && shouldShowCreateActions);
-  const shouldShowDeleteActions = Boolean(item) && !isTabMenu;
-  const shouldShowDividerBeforeDelete = shouldShowDeleteActions;
-  const icon = (Icon: LucideIcon) => <Icon size={14} />;
-
-  return (
-    <UiContextMenu x={pos.x} y={pos.y} items={((): UiMenuItem[] => {
-      const entries: UiMenuItem[] = [];
-      if (item && item.isDirectory) {
-        entries.push({ label: t('在新标签页打开'), icon: icon(FolderOpen), onSelect: onOpenInNewTab });
-      }
-      if (isTabMenu && canTogglePinned && !isSystemPinned) {
-        entries.push({ label: isPinned ? t('取消固定') : t('固定'), icon: icon(Pin), onSelect: onTogglePinned });
-      }
-      if (canCloseTab) {
-        entries.push({ label: t('关闭标签'), icon: icon(X), onSelect: onCloseTab });
-      }
-      if (item) {
-        entries.push({ label: t('复制路径'), icon: icon(Copy), onSelect: onCopyPath });
-      }
-      if (item && !isTabMenu) {
-        entries.push({
-          label: `${clipboardActionArrow === '<-' ? `${clipboardActionArrow} ${t('复制')}` : `${t('复制')}${clipboardActionArrow ? ` ${clipboardActionArrow}` : ''}`}${clipboardItemCount > 1 ? ` (${clipboardItemCount}${t('项')})` : ''}`,
-          icon: icon(Copy),
-          onSelect: onCopyItem,
-        });
-      }
-      if (item && !isTabMenu) {
-        entries.push({
-          label: `${clipboardActionArrow === '<-' ? `${clipboardActionArrow} ${t('剪切')}` : `${t('剪切')}${clipboardActionArrow ? ` ${clipboardActionArrow}` : ''}`}${clipboardItemCount > 1 ? ` (${clipboardItemCount}${t('项')})` : ''}`,
-          icon: icon(Scissors),
-          onSelect: onCutItem,
-        });
-      }
-      if (!isTabMenu && canPaste) {
-        entries.push({ label: t('粘贴'), icon: icon(ClipboardPaste), onSelect: onPaste });
-      }
-      if (item && !item.isDirectory && isEditable(item.name)) {
-        entries.push({ label: t('编辑'), icon: icon(SquarePen), onSelect: onEdit });
-      }
-      if (item && !item.isDirectory) {
-        entries.push({ label: t('系统编辑器打开'), icon: icon(MonitorSmartphone), onSelect: onOpenSystemEditor });
-      }
-      if (item && !item.isDirectory) {
-        entries.push({ label: t('指定编辑器打开'), icon: icon(PencilLine), onSelect: onOpenWithEditor });
-      }
-      if (item) {
-        entries.push({ label: item.isDirectory ? t('下载文件夹到本地') : t('下载到本地'), icon: icon(Download), onSelect: onDownload });
-      }
-      if (item) {
-        entries.push({ label: t('压缩 (tar.gz)'), icon: icon(Archive), onSelect: onCompress });
-      }
-      if (item && !item.isDirectory && isArchive(item.name)) {
-        entries.push({ label: t('解压'), icon: icon(FileArchive), onSelect: onUncompress });
-      }
-      if (item && (!isTabMenu || !isSystemPinned)) {
-        entries.push({ label: isTabMenu ? t('重命名标签标题') : t('重命名'), icon: icon(PenLine), onSelect: onRename });
-      }
-      if (item) {
-        entries.push({ label: t('修改权限'), icon: icon(Lock), onSelect: onChmod });
-      }
-      if (shouldShowDividerBeforeCreate) {
-        entries.push('separator');
-      }
-      if (shouldShowCreateActions) {
-        entries.push({ label: t('新建文件'), icon: icon(FilePlus), onSelect: onNewFile });
-        entries.push({ label: t('新建文件夹'), icon: icon(FolderPlus), onSelect: onMkdir });
-      }
-      if (shouldShowDividerBeforeDelete) {
-        entries.push('separator');
-      }
-      if (shouldShowDeleteActions) {
-        entries.push({
-          label: `${t('删除')}${deleteItemCount > 1 ? ` (${deleteItemCount}${t('项')})` : ''}`,
-          icon: icon(Trash2),
-          danger: true,
-          onSelect: onDelete,
-        });
-      }
-      return entries;
-    })()} onClose={onClose} zIndex={Z.MENU} />
-  );
-}
+import ChmodDialog from './filemanager/ChmodDialog.tsx';
+import RenameInput from './filemanager/RenameInput.tsx';
+import ContextMenu from './filemanager/ContextMenu.tsx';
+import type {
+  FileManagerDownloadConflict,
+  FileManagerDualPaneDragItem,
+  FileManagerDualPaneDragPayload,
+  FileManagerFileLike,
+  FileManagerPaneStateLike,
+  FileManagerProps,
+  FileManagerTabDropIndicator,
+  FileManagerVirtualRowEntry,
+  FileManagerWorkspace,
+  FileManagerWorkspaceTab,
+  FileManagerChmodTarget,
+  FileManagerFileItem,
+  ContextMenuState,
+  LoadDirOptions,
+  PaneSelectionRestore,
+  SyncTabOverrides,
+} from './filemanager/fileManagerTypes.ts';
+import type { LooseT } from './filemanager/fileManagerTypes.ts';
 
 export default function FileManager({ sessionId, sessionGroupId = sessionId, addToast, isActive = true, initialPath = '' }: FileManagerProps) {
   const { t } = useTranslation();
